Ensure that shutdown_all takes potential ssl values into account
The code to handled the shutdown-all request (when Notebook receives
`SIGTERM`) was not connecting to the _Gateway_ server with the required
`KG_` values that convey information necessary for proper SSL behavior.

The `shutdown_all()` method doesn't use the async `fetch_kg()` method
which contains the necessary code to convey SSL information.  This
change moves that code to a method so that it can be leveraged by other
methods that cannot use `fetch_kg()` to connect to the gateway.