Convey working directory of notebook
When communicating with a Gateway that launches container-based kernels,
it is desirable to mirror the notebook's working directory in cases where
the container is sharing the notebook directory's volume.  This is a
common configuration in Jupyter Hub installations running on Kubernetes
where the Notebook/Lab instance enables the mounting of the user's home
directory and those same mounts are used via the Enterprise Gateway
kernel images.  This change adds `KERNEL_WORKING_DIR` to the json body
of the kernel creation request which enables the receiving Gateway to
potentially act on that value when launching kernel images.