Add support for retrieving kernelspec resources
A long-standing TODO item has been to handle the retrieval of resource files
corresponding to the kernelspecs.  This change forwards such a request to
the configured gateway if the corresponding kernelspec uses resources files
(logos, kernel.js or kernel.css) and statisfies the request from the Notebook
client.

This change relies an appropriate Jupyter Kernel Gateway to retrieve the resource.
If used against an older version of JKG, no image is displayed, etc.